Add quests/quest givers to wiki DB.

See merge request Meridian59/Meridian59!1170

+++ blakserv/message.h
@@ -27,8 +27,8 @@ typedef struct message_struct
    char *handler;
    int dstr_id;
    int trace_session_id;
-   int untimed_call_count;
-   int timed_call_count;
+   UINT64 untimed_call_count;
+   UINT64 timed_call_count;
    struct message_struct *propagate_message;
    struct class_struct *propagate_class;
    double total_call_time;
